+++ TS.EasyStockManager.Web/Data/Entity/StoreStock.cs
@@ -8,7 +8,7 @@ namespace Aby.StockManager.Data.Entity
     {
         public int StoreId { get; set; }
         public int ProductId { get; set; }
-        public decimal Stock { get; set; }
+        public int Stock { get; set; }
         public virtual Product Product { get; set; }
         public virtual Store Store { get; set; }
     }
