+++ TS.EasyStockManager.Web/Model/Domain/ProductDTO.cs
@@ -10,11 +10,10 @@ namespace Aby.StockManager.Model.Domain
         public string Barcode { get; set; }
         public string Description { get; set; }
         public string Image { get; set; }
-        public decimal? Price { get; set; }
+        public double? Price { get; set; }
         public int? CategoryId { get; set; }
         public string CategoryName { get; set; }
         public int? UnitOfMeasureId { get; set; }
         public string UnitOfMeasureName { get; set; }
-
     }
 }
