+++ TS.EasyStockManager.Web/Model/Domain/StoreStockDTO.cs
@@ -8,7 +8,7 @@ namespace Aby.StockManager.Model.Domain
     {
         public int? StoreId { get; set; }
         public int? ProductId { get; set; }
-        public decimal? Stock { get; set; }
+        public double? Stock { get; set; }
         public string StoreName { get; set; }
         public string ProductName { get; set; }
         public string Barcode { get; set; }
